+++ src/components/Container/index.ts
@@ -1,0 +1,2 @@
+export { Container } from "./Container";
+export type { ContainerProps, ContainerSize, ContainerAlign } from "./Container.types";