+++ src/components/index.ts
@@ -8,6 +8,7 @@ export { Box } from "./Box/Box";
 export type { BoxProps } from "./Box/Box.types";
 export * from "./Flex";
 export * from "./Grid";
+export * from "./Container";
 
 // Hooks
 export * from "../hooks/useButtonBase";